Pull up question from the db in the exercise space
